1. Zmiany w plikach html oraz style.css. 2. Dodatkowo zmianiono typ zmiennej wheelSize z int na float.

+++ src/main/java/pl/edu/wszib/jwd/onlinestore/data/entities/BicycleEntity.java
@@ -39,7 +39,7 @@ public class BicycleEntity {
     private int frameSize;
 
     @Column(name = "wheelSize")
-    private int wheelSize;
+    private float wheelSize;
 
     @Column(name = "numberOfGears")
     private int numberOfGears;
@@ -120,11 +120,11 @@ public class BicycleEntity {
         this.frameSize = frameSize;
     }
 
-    public int getWheelSize() {
+    public float getWheelSize() {
         return wheelSize;
     }
 
-    public void setWheelSize(int wheelSize) {
+    public void setWheelSize(float wheelSize) {
         this.wheelSize = wheelSize;
     }
